+++ scripts/library/rollup.config.iife.mjs
@@ -1,12 +1,15 @@
+import fs from "fs";
+import path from "path";
+import {fileURLToPath} from "url";
+
 import babel from "@rollup/plugin-babel";
-import stripCode from "rollup-plugin-strip-code";
 import replace from "@rollup/plugin-replace";
+import license from "rollup-plugin-license";
+import stripCode from "rollup-plugin-strip-code";
+import versionInjector from "rollup-plugin-version-injector";
 
-const fs = require("fs");
-const license = require("rollup-plugin-license");
-const versionInjector = require("rollup-plugin-version-injector");
-
-const BANNER = fs.readFileSync(__dirname + "/../../BANNER.txt", "utf8") + "\n\n\n";
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const BANNER = fs.readFileSync(path.join(__dirname, "..", "..", "BANNER.txt"), "utf8") + "\n\n\n";
 
 export default {
 
@@ -27,7 +30,7 @@ export default {
       start_comment: "START-ESM",
       end_comment: "END-ESM"
     }),
-    babel(),
+    babel({babelHelpers: "bundled"}),
     license({
       banner: BANNER
     })